+++ examples/13-enable-sessionkey-module.ts
@@ -25,7 +25,6 @@ async function main() {
   const token = '0x1c7D4B196Cb0C7B01d743Fbc6116a902379C7238';
   const functionSelector = '0xa9059cbb';
   const spendingLimit = '100000';
-  const validAfter = new Date().getTime();
   const validUntil = new Date().getTime() + 24 * 60 * 60 * 1000;
 
   // get instance  of SessionKeyValidator
@@ -38,7 +37,6 @@ async function main() {
     token,
     functionSelector,
     spendingLimit,
-    validAfter,
     validUntil,
     KeyStore.AWS
   );
@@ -58,6 +56,9 @@ async function main() {
 
   const sessionKeys = await sessionKeyModule.getAssociatedSessionKeys();
   console.log('\x1b[33m%s\x1b[0m', `AssociatedSessionKeys: `, sessionKeys);
+
+  const sessionData = await sessionKeyModule.sessionData(response.sessionKey);
+  console.log('\x1b[33m%s\x1b[0m', `SessionData: `, sessionData);
 }
 
 main()