+++ src/sdk/SessionKeyValidator/SessionKeyValidator.ts
@@ -1,10 +1,12 @@
 import { ERC20SessionKeyValidator__factory } from "../contracts/factories/src/ERC7579/modules/validator";
-import { BigNumber, providers } from "ethers";
+import { providers } from "ethers";
 import { ModularSdk } from "../sdk";
 import { KeyStore, PERMISSIONS_URL } from "./constants";
-import { SessionKeyResponse, GenerateSessionKeyResponse, GetNonceResponse, GetSessionKeyResponse, DeleteSessionKeyResponse } from "./interfaces";
+import { SessionKeyResponse, GenerateSessionKeyResponse, GetNonceResponse, GetSessionKeyResponse, DeleteSessionKeyResponse, SessionData } from "./interfaces";
 import { BundlerProvider } from "../bundler";
 import { DEFAULT_ERC20_SESSION_KEY_VALIDATOR_ADDRESS, Networks } from "../network/constants";
+import { UserOperation, deepHexlify } from "../common";
+import { resolveProperties } from "ethers/lib/utils";
 
 export class SessionKeyValidator {
     private modularSdk: ModularSdk;
@@ -21,7 +23,6 @@ export class SessionKeyValidator {
         token: string,
         functionSelector: string,
         spendingLimit: string,
-        validAfter: number,
         validUntil: number,
         keyStore?: KeyStore,
     ): Promise<SessionKeyResponse> {
@@ -38,7 +39,6 @@ export class SessionKeyValidator {
                 token,
                 functionSelector,
                 spendingLimit,
-                validAfter,
                 validUntil,
                 apiKey,
                 false,
@@ -56,15 +56,20 @@ export class SessionKeyValidator {
 
             await this.modularSdk.addUserOpsToBatch({ to: erc20SessionKeyValidator, data: enableSessionKeyData });
 
-            const op = await this.modularSdk.estimate({
-                key: BigNumber.from(erc20SessionKeyValidator)
-            });
+            try {
+                const op = await this.modularSdk.estimate();
 
-            const uoHash = await this.modularSdk.send(op)
+                const uoHash = await this.modularSdk.send(op);
 
-            return {
-                userOpHash: uoHash,
-                sessionKey: data.sessionKey,
+                if (!uoHash) await this.deleteSessionKey(account, chainId, apiKey, data.sessionKey);
+
+                return {
+                    userOpHash: uoHash,
+                    sessionKey: data.sessionKey,
+                }
+            } catch (error) {
+                await this.deleteSessionKey(account, chainId, apiKey, data.sessionKey);
+                throw error;
             }
         } catch (error) {
             throw error;
@@ -75,8 +80,8 @@ export class SessionKeyValidator {
         token: string,
         functionSelector: string,
         spendingLimit: string,
-        validAfter: number,
         validUntil: number,
+        oldSessionKey: string,
         keyStore?: KeyStore,
     ): Promise<SessionKeyResponse> {
         try {
@@ -92,11 +97,11 @@ export class SessionKeyValidator {
                 token,
                 functionSelector,
                 spendingLimit,
-                validAfter,
                 validUntil,
                 apiKey,
                 true,
                 keyStore ? keyStore : null,
+                oldSessionKey,
             )
 
             const erc20SessionKeyValidatorContract = ERC20SessionKeyValidator__factory.connect(
@@ -112,22 +117,32 @@ export class SessionKeyValidator {
 
             await this.modularSdk.addUserOpsToBatch({ to: erc20SessionKeyValidator, data: rotateSessionKeyData });
 
-            const op = await this.modularSdk.estimate({
-                key: BigNumber.from(erc20SessionKeyValidator)
-            });
+            try {
+                const op = await this.modularSdk.estimate();
 
-            const uoHash = await this.modularSdk.send(op);
+                const uoHash = await this.modularSdk.send(op);
 
-            return {
-                userOpHash: uoHash,
-                sessionKey: data.sessionKey,
+                if (uoHash) {
+                    await this.deleteSessionKey(account, chainId, apiKey, data.oldSessionKey);
+                }
+                else {
+                    await this.deleteSessionKey(account, chainId, apiKey, data.sessionKey);
+                }
+
+                return {
+                    userOpHash: uoHash,
+                    sessionKey: data.sessionKey,
+                }
+            } catch (error) {
+                await this.deleteSessionKey(account, chainId, apiKey, data.sessionKey);
+                throw error;
             }
         } catch (error) {
             throw error;
         }
     }
 
-    async disableSessionKey(): Promise<SessionKeyResponse> {
+    async disableSessionKey(sessionKey: string): Promise<SessionKeyResponse> {
         try {
             const account = await this.modularSdk.getCounterFactualAddress();
             const erc20SessionKeyValidator = await this.getERC20SessionKeyValidator();
@@ -139,6 +154,7 @@ export class SessionKeyValidator {
                 account,
                 chainId,
                 apiKey,
+                sessionKey,
             )
 
             const erc20SessionKeyValidatorContract = ERC20SessionKeyValidator__factory.connect(
@@ -154,14 +170,12 @@ export class SessionKeyValidator {
 
             await this.modularSdk.addUserOpsToBatch({ to: erc20SessionKeyValidator, data: disableSessionKeyData });
 
-            const op = await this.modularSdk.estimate({
-                key: BigNumber.from(erc20SessionKeyValidator)
-            });
+            const op = await this.modularSdk.estimate();
 
             const uoHash = await this.modularSdk.send(op);
 
             if (uoHash) {
-                await this.deleteSessionKey(account, chainId, apiKey);
+                await this.deleteSessionKey(account, chainId, apiKey, sessionKey);
             }
 
             return {
@@ -173,7 +187,28 @@ export class SessionKeyValidator {
         }
     }
 
-    async getNonce(): Promise<GetNonceResponse> {
+    async signUserOpWithSessionKey(sessionKey: string, userOp: UserOperation): Promise<UserOperation> {
+        try {
+            const account = await this.modularSdk.getCounterFactualAddress();
+            const chainId = await this.getChainId();
+            const apiKeyMatch = this.provider.connection.url.match(/api-key=([^&]+)/);
+            const apiKey = apiKeyMatch ? apiKeyMatch[1] : null;
+
+            const data: UserOperation = await this.getSignUserOp(
+                account,
+                chainId,
+                apiKey,
+                sessionKey,
+                userOp,
+            )
+
+            return data;
+        } catch (error) {
+            throw error;
+        }
+    }
+
+    async getNonce(sessionKey: string): Promise<GetNonceResponse> {
         try {
             const account = await this.modularSdk.getCounterFactualAddress();
             const chainId = await this.getChainId();
@@ -184,6 +219,7 @@ export class SessionKeyValidator {
                 account,
                 chainId,
                 apiKey,
+                sessionKey,
             )
 
             return data;
@@ -203,6 +239,30 @@ export class SessionKeyValidator {
         );
 
         return await erc20SessionKeyValidatorContract.callStatic.getAssociatedSessionKeys({ from: account });
+    }
+
+    async sessionData(sessionKey: string): Promise<SessionData> {
+        const account = await this.modularSdk.getCounterFactualAddress();
+
+        const erc20SessionKeyValidator = await this.getERC20SessionKeyValidator();
+
+        const erc20SessionKeyValidatorContract = ERC20SessionKeyValidator__factory.connect(
+            erc20SessionKeyValidator,
+            this.provider
+        );
+
+        const data = await erc20SessionKeyValidatorContract.callStatic.sessionData(sessionKey, account);
+
+        const { token, funcSelector, spendingLimit, validAfter, validUntil, live } = data;
+
+        return {
+            token,
+            funcSelector,
+            spendingLimit,
+            validAfter,
+            validUntil,
+            live
+        }
     }
 
     private async getERC20SessionKeyValidator(): Promise<string> {
@@ -229,11 +289,11 @@ export class SessionKeyValidator {
         token: string,
         functionSelector: string,
         spendingLimit: string,
-        validAfter: number,
         validUntil: number,
         apiKey: string,
         rotateKey: boolean,
         keyStore: KeyStore,
+        oldSessionKey?: string,
     ): Promise<GenerateSessionKeyResponse> {
         let response = null;
         try {
@@ -248,8 +308,8 @@ export class SessionKeyValidator {
                 token,
                 functionSelector,
                 spendingLimit,
-                validAfter,
                 validUntil,
+                oldSessionKey,
             };
 
             response = await fetch(url, {
@@ -277,11 +337,12 @@ export class SessionKeyValidator {
         account: string,
         chainId: number,
         apiKey: string,
+        sessionKey: string,
     ): Promise<GetSessionKeyResponse> {
         let response = null;
 
         try {
-            let url = `${PERMISSIONS_URL}/account/getSessionKey?account=${account}&chainId=${chainId}`;
+            let url = `${PERMISSIONS_URL}/account/getSessionKey?account=${account}&chainId=${chainId}&sessionKey=${sessionKey}`;
             if (apiKey) url += `&apiKey=${apiKey}`;
 
             response = await fetch(url, {
@@ -308,10 +369,11 @@ export class SessionKeyValidator {
         account: string,
         chainId: number,
         apiKey: string,
+        sessionKey: string,
     ): Promise<DeleteSessionKeyResponse> {
         let response = null;
         try {
-            let url = `${PERMISSIONS_URL}/account/deleteSessionKey?account=${account}&chainId=${chainId}`;
+            let url = `${PERMISSIONS_URL}/account/deleteSessionKey?account=${account}&chainId=${chainId}&sessionKey=${sessionKey}`;
             if (apiKey) url += `&apiKey=${apiKey}`;
 
             response = await fetch(url, {
@@ -334,15 +396,50 @@ export class SessionKeyValidator {
         }
     }
 
+    private async getSignUserOp(
+        account: string,
+        chainId: number,
+        apiKey: string,
+        sessionKey: string,
+        userOp: UserOperation,
+    ): Promise<UserOperation> {
+        let response = null;
+
+        try {
+            let url = `${PERMISSIONS_URL}/account/signUserOp?account=${account}&chainId=${chainId}&sessionKey=${sessionKey}`;
+            if (apiKey) url += `&apiKey=${apiKey}`;
+
+            response = await fetch(url, {
+                method: 'POST',
+                headers: {
+                    'Accept': 'application/json',
+                    'Content-Type': 'application/json',
+                },
+                body: JSON.stringify(deepHexlify(await resolveProperties(userOp))),
+            });
+
+            if (response.status === 200) {
+                const responseJson = await response.json();
+                return responseJson
+            } else {
+                const responseJson = await response.json();
+                throw new Error(responseJson.message)
+            }
+        } catch (err) {
+            throw new Error(err.message)
+        }
+    }
+
     private async getNonceData(
         account: string,
         chainId: number,
         apiKey: string,
+        sessionKey: string,
     ): Promise<GetNonceResponse> {
         let response = null;
 
         try {
-            let url = `${PERMISSIONS_URL}/account/getNonce?account=${account}&chainId=${chainId}`;
+            let url = `${PERMISSIONS_URL}/account/getNonce?account=${account}&chainId=${chainId}&sessionKey=${sessionKey}`;
             if (apiKey) url += `&apiKey=${apiKey}`;
 
             response = await fetch(url, {